feat: add Supabase delete method to deleteFile
- add delete method to delete files from Supabase file storage
- add error handling for Supabase file storage deletion

+++ controllers/fileController.cjs
@@ -129,6 +129,13 @@ async function updateFileFolder(req, res, next) {
 async function deleteFile(req, res, next) {
   try {
     const fileId = Number(req.params.fileId);
+    const file = await fileRepository.getFileById(fileId);
+
+    const { data, error } = await supabase.storage.from(file.bucket).remove([`${file.path}`]);
+
+    if (error) return next(error);
+    if (!data || data.length === 0) return next({ statusCode: 404, message: "File not found" });
+
     await fileRepository.deleteFileById(fileId);
 
     const path = req.body.folderId ? `/folder/${req.body.folderId}` : "/";
